requirements.txt
ApiRequestManager  1.0.5
async-generator    1.10
atomicwrites       1.4.1
attrs              22.2.0
beautifulsoup4     4.11.2
certifi            2022.12.7
cffi               1.15.1
charset-normalizer 3.0.1
colorama           0.4.6
django-bublik      0.1.4
exceptiongroup     1.1.1
Faker              17.3.0
h11                0.14.0
idna               3.4
iniconfig          2.0.0
jsons              1.6.3
outcome            1.2.0
packaging          23.0
password           0.2
pf                 0.1.1
Pillow             9.4.0
pip                22.3.1
pluggy             1.0.0
py                 1.11.0
pycparser          2.21
pydantic           1.10.7
PySocks            1.7.1
pyTelegramBotAPI   4.10.0
pytest             7.3.1
pytest-base-url    2.0.0
pytest-html        3.2.0
pytest-metadata    2.0.4
pytest-selenium    4.0.0
pytest-variables   2.0.0
python-dateutil    2.8.2
python-dotenv      1.0.0
requests           2.28.2
requests-toolbelt  0.10.1
selenium           4.9.0
setuptools         65.5.0
six                1.16.0
sniffio            1.3.0
sortedcontainers   2.4.0
soupsieve          2.4
telebot            0.0.5
tenacity           6.3.1
termcolor          2.3.0
toml               0.10.2
trio               0.22.0
trio-websocket     0.10.2
typing_extensions  4.5.0
typish             1.9.3
urllib3            1.26.14
vnv                2.1
wsproto            1.2.0
